Move logcat pubsub module in services directory.
